// lib/generated/BajsRideHour.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type BajsRideHour = { 
/**
 * Zagreb local hour, 0-23.
 */
hour: number, 
/**
 * Rides started in this hour per day observed, averaged.
 */
avgStarts: number, avgReturns: number, 
/**
 * How many well-observed days went into the average.
 */
days: number, };